Reorganised code to separate static models from database-persisted ones
