+++ app/controllers/api/v1/api_controller.rb
@@ -27,7 +27,7 @@ module Api
       private
 
       def load_section
-        @section = Section.find_by_slug(params[:section_slug])
+        @section = Static::Section.find_by_slug(params[:section_slug])
         render json: {}, status: :not_found and return unless @section
       end
     end
