+++ app/controllers/api/v1/sections_controller.rb
@@ -4,20 +4,18 @@ module Api
       before_action :load_section, only: [:show]
 
       def index
-        @sections = Section.all.map do |section|
-          section.to_hash(section_includes)
-        end
+        @sections = GetAllSections.new.call(section_includes)
         render json: @sections
       end
 
       def show
-        render json: @section.to_hash(section_includes)
+        render json: @section
       end
 
       private
 
       def load_section
-        @section = Section.find_by_slug(params[:slug])
+        @section = GetSection.new.call(params[:slug], section_includes)
         render json: {}, status: :not_found and return unless @section
       end
