+++ app/services/get_all_categories.rb
@@ -1,0 +1,7 @@
+class GetAllCategories
+  def call(section, category_includes)
+    section.categories.map do |category|
+      category.to_hash(category_includes)
+    end
+  end
+end
