+++ app/services/get_all_sections.rb
@@ -1,0 +1,7 @@
+class GetAllSections
+  def call(section_includes)
+    Static::Section.all.map do |section|
+      section.to_hash(section_includes)
+    end
+  end
+end
