+++ app/services/get_category.rb
@@ -1,0 +1,7 @@
+class GetCategory
+  def call(section, slug, category_includes)
+    category = section.find_category_by_slug(slug)
+    return nil unless category
+    category.to_hash(category_includes)
+  end
+end
